laird: som60: Boot kernel from UBI
Provide helper functions to format and set the boot volumes to allow
booting, the volume number of the rootfs is needed to be set as the name
of the volume cannot be used in the bootargs.
This volume number ubi format dependant so we include the format that
matches.

Bug 12760