laird: add ability to store a fs key in some memory
The memory used should be outside of the stack and heap and on the
internal CPU memory, not main memory.
The structure of the control dtb is in the same form as the other
encryption information
encryption {
	laird,fs-key = <0x00 ..... >;
};

Which needs to be injected to the dtb prior to encryption and u-boot.itb
creation.
This work is coupled with the kernel laird fs injector work, and is useless
without it.

Bug 12815