NAND partition updates (continued)
Reduce UBoot partition by 1 block to 1MB, and add that block to UBI partition.
This is the preferred UBoot size, and the single block can be used in the UBI partition now that we plan to implement a permanent volume separate from the a/b sides.
UBI partitions coming in a later commit.

Bug: 14492
(cherry picked from commit 9898163fd5d256a15c607500b135c23c15ca979f)