NAND: Blocks reserved for BBT incorrectly reported as Bad
Blocks that are used to store the BBT are marked as Reserved and reported as Bad
in order to protect them from modification.  Since they are reported as Bad, the
bad block check in various places indicates them as bad to an external user.  Add
another check to see if they are really marked as Reserved and report the correct
status to the user.

Bug: 15373