uboot: Fixed random boot failure
Boot fails with the message:
"SMC/TWI/QSPI boot device not supported!"

Code tries to examine R4 CPU register left over by ROMBoot.
Sometimes content of this register is different from expected.

From original commit that introduced the functionality, this
autodetect code is only useful when figuring out which MMC
to boot from (in our case MMC0 is the only choice) the rest of
the functionality is redundant with u-boot configurations.

Bug 19104