uboot: wb45/wb50 legacy: Re-enable crc32 command and revert prompt
The crc32 command is currently required for WBx3 to program the WB45/50 legacy
targets and is not present when 'CONFIG_CMD_MEMORY' is disabled. Therefore, this
change re-enables the config.

Also, the WB45/50 legacy U-Boot prompt changed to the default "=> ", and this
commit reverts it back to the previous value of "U-Boot> ".

Bug 23401